ajout de réservation et de auth.service pour les pages du menu et de la carte

+++ src/app/Components/carte/carte.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CarteService } from 'src/app/Services/carte.service';
 import { Carte } from 'src/app/Models/carte';
+import { AuthService } from 'src/app/Services/auth.service';
 
 @Component({
   selector: 'app-carte',
@@ -12,23 +13,29 @@ export class CarteComponent implements OnInit {
   cartes: Carte[] = [];
   isAdmin = false;
 
-  constructor(private carteService: CarteService) { }
+  constructor(private carteService: CarteService,private authService: AuthService) { }
 
   ngOnInit(): void {
     this.cartes = this.carteService.getListe();
   }
 
   canEdit(): boolean {
-    return this.isAdmin === true;
+    return this.authService.isAdmin();
   }
 
   updateCarte(id: number, nom: string, prix: number): void {
-    this.carteService.updateCarte(id, { nom, prix });
+    if (this.authService.isAuthenticated()) {
+      this.carteService.updateCarte(id, { nom, prix });
+    }
   }
 
+
   deleteCarte(id: number): void {
-    this.carteService.deleteCarte(id);
-    this.cartes = this.carteService.getListe();
+    if (this.authService.isAuthenticated()) {
+      this.carteService.deleteCarte(id);
+      this.cartes = this.carteService.getListe();
+    }
   }
+
 
 }
